+++ src/runner.mjs
@@ -1,0 +1,7 @@
+#!/usr/bin/env node
+
+import { checkCliArgs } from './cli.mjs'
+import { run } from './core.mjs'
+
+checkCliArgs()
+run()