+++ src/setup.mjs
@@ -1,40 +1,12 @@
 import path from 'path'
 import fs from 'fs'
+import { getConfigName } from './cli.mjs'
 
-const getCustomArgFromArgs = (args, customArgPrefix) => {
-  // pick custom arg that contains custom prefix
-  const customArg = args.find((arg) => arg.includes(customArgPrefix)) ?? ''
-  const parsedCustomArg = customArg.split(customArgPrefix)
-  const customArgValue = parsedCustomArg[1]
-  return customArgValue
-}
-
-const readConfigJSONFile = (args) => {
-  if (getConfigName(args)) {
-    const file = fs.readFileSync(
-      path.resolve(process.cwd(), getConfigName(args))
-    )
+export const readConfigJSONFile = (args) => {
+  if (getConfigName()) {
+    const file = fs.readFileSync(path.resolve(process.cwd(), getConfigName()))
     return JSON.parse(file)
   }
-}
-
-const getConfigName = (args) => {
-  const customArgPrefix = '--config='
-  return getCustomArgFromArgs(args, customArgPrefix) || ''
-}
-
-export const getTestNameFromArgs = (args) => {
-  const customArgPrefix = '--test='
-  return getCustomArgFromArgs(args, customArgPrefix) || ''
-}
-
-export const getTestFolderFromArgs = (args) => {
-  const customArgPrefix = '--test-folder='
-  return getCustomArgFromArgs(args, customArgPrefix) || ''
-}
-
-export const getConfigFileFromArgs = (args) => {
-  return readConfigJSONFile(args)
 }
 
 const getAllFilePaths = (dir) => {