+++ CHANGELOG.md
@@ -14,6 +14,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   Option timeout (in ms) for specifying how long to wait before aborting.
   The default timeout is 5 seconds. 
     `test('description', { timeout: 2000 }, () => {})`
+- Assertions:
+   - `.toBeNull()`
+   - `.notToBeNull()`
+   - `.toBeUndefined()`
+   - `.toBeNaN()`
+   - `.toBeGreaterThan()`
+   - `.toBeLessThan()`
+
 
 ### Fixed
 - Fixed EOL for assertions