+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Logic for handling and erroring args 
   https://github.com/VadimNastoyashchy/real-test-js/issues/78
 
+### Changed
+- Refactored cli, setup, .mjs
+
 ## 1.8.0 - 2024-01-17
 ### Added
 - Js dock